+++ src/RootSuite.ts
@@ -1,14 +1,14 @@
 import * as vscode from 'vscode';
 import { TestInfo } from 'vscode-test-adapter-api';
-import { TestExecutableInfo } from './TestExecutableInfo';
-import { AbstractTestSuiteInfoBase } from './AbstractTestSuiteInfoBase';
-import { AbstractRunnableTestSuiteInfo } from './AbstractRunnableTestSuiteInfo';
-import { AbstractTestInfo } from './AbstractTestInfo';
+import { Executable } from './Executable';
+import { AbstractSuit } from './AbstractSuit';
+import { AbstractRunnableSuite } from './AbstractRunnableSuite';
+import { AbstractTest } from './AbstractTest';
 import { SharedVariables } from './SharedVariables';
 
-export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vscode.Disposable {
-  public readonly children: AbstractRunnableTestSuiteInfo[] = [];
-  private _executables: TestExecutableInfo[] = [];
+export class RootSuite extends AbstractSuit implements vscode.Disposable {
+  public readonly children: AbstractRunnableSuite[] = [];
+  private _executables: Executable[] = [];
 
   public constructor(id: string | undefined, shared: SharedVariables) {
     super(shared, 'Catch2 and Google tests', undefined, id);
@@ -22,7 +22,7 @@ export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vsco
     this.children.forEach(c => c.cancel());
   }
 
-  public load(executables: TestExecutableInfo[]): Promise<void> {
+  public load(executables: Executable[]): Promise<void> {
     this._executables.forEach(e => e.dispose());
 
     this._executables = executables;
@@ -69,16 +69,16 @@ export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vsco
       });
   }
 
-  public hasChild(suite: AbstractRunnableTestSuiteInfo): boolean {
+  public hasChild(suite: AbstractRunnableSuite): boolean {
     return this.children.indexOf(suite) != -1;
   }
 
-  public insertChild(suite: AbstractRunnableTestSuiteInfo, uniquifyLabels: boolean): boolean {
+  public insertChild(suite: AbstractRunnableSuite, uniquifyLabels: boolean): boolean {
     if (this.hasChild(suite)) return false;
 
     {
       // we want to filter the situation when 2 patterns match the same file
-      const other = this.children.find((s: AbstractRunnableTestSuiteInfo) => {
+      const other = this.children.find((s: AbstractRunnableSuite) => {
         return suite.execInfo.path == s.execInfo.path;
       });
       if (other) {
@@ -102,7 +102,7 @@ export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vsco
     return true;
   }
 
-  public removeChild(child: AbstractRunnableTestSuiteInfo): boolean {
+  public removeChild(child: AbstractRunnableSuite): boolean {
     const i = this.children.findIndex(val => val.id == child.id);
     if (i != -1) {
       this.children.splice(i, 1);
@@ -113,7 +113,7 @@ export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vsco
   }
 
   public uniquifySuiteLabels(): void {
-    const uniqueNames = new Map<string /* name */, AbstractRunnableTestSuiteInfo[]>();
+    const uniqueNames = new Map<string /* name */, AbstractRunnableSuite[]>();
 
     for (const suite of this.children) {
       const suites = uniqueNames.get(suite.origLabel);
@@ -134,9 +134,7 @@ export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vsco
     }
   }
 
-  public findRouteToTestInfo(
-    pred: (v: AbstractTestInfo) => boolean,
-  ): [AbstractTestSuiteInfoBase[], AbstractTestInfo | undefined] {
+  public findRouteToTestInfo(pred: (v: AbstractTest) => boolean): [AbstractSuit[], AbstractTest | undefined] {
     for (let i = 0; i < this.children.length; ++i) {
       const found = this.children[i].findRouteToTestInfo(pred);
       if (found[1] !== undefined) {
@@ -146,7 +144,7 @@ export class RootTestSuiteInfo extends AbstractTestSuiteInfoBase implements vsco
     return [[], undefined];
   }
 
-  public findRouteToTest(idOrInfo: string | TestInfo): [AbstractTestSuiteInfoBase[], AbstractTestInfo | undefined] {
+  public findRouteToTest(idOrInfo: string | TestInfo): [AbstractSuit[], AbstractTest | undefined] {
     if (typeof idOrInfo === 'string') return this.findRouteToTestInfo(x => x.id === idOrInfo);
     else return this.findRouteToTestInfo(x => x === idOrInfo);
   }
