+++ src/RunnableSuiteProperties.ts
@@ -1,7 +1,7 @@
 import * as c2fs from './FSWrapper';
-import { TestExecutableInfoFrameworkSpecific } from './TestExecutableInfo';
+import { TestExecutableInfoFrameworkSpecific } from './Executable';
 
-export class RunnableTestSuiteProperties {
+export class RunnableSuiteProperties {
   public constructor(
     public readonly path: string,
     public readonly options: c2fs.SpawnOptions,
