+++ src/SharedVariables.ts
@@ -1,7 +1,7 @@
 import { LoggerWrapper } from './LoggerWrapper';
 import * as vscode from 'vscode';
 import { TestRunStartedEvent, TestRunFinishedEvent, TestSuiteEvent, TestEvent } from 'vscode-test-adapter-api';
-import { AbstractRunnableTestSuiteInfo } from './AbstractRunnableTestSuiteInfo';
+import { AbstractRunnableSuite } from './AbstractRunnableSuite';
 import { TaskPool } from './TaskPool';
 
 type TestStateEmitterType = vscode.EventEmitter<
@@ -18,7 +18,7 @@ export class SharedVariables implements vscode.Disposable {
     public readonly testStatesEmitter: TestStateEmitterType,
     public readonly loadWithTaskEmitter: vscode.EventEmitter<() => void | PromiseLike<void>>,
     public readonly sendTestEventEmitter: vscode.EventEmitter<TestEvent[]>,
-    public readonly retire: vscode.EventEmitter<AbstractRunnableTestSuiteInfo[]>,
+    public readonly retire: vscode.EventEmitter<AbstractRunnableSuite[]>,
     public rngSeed: string | number | null,
     public execWatchTimeout: number,
     public retireDebounceTime: number,
