+++ src/framework/Catch2Suite.ts
@@ -5,66 +5,116 @@ import * as xml2js from 'xml2js';
 import * as pathlib from 'path';
 
 import * as c2fs from '../FSWrapper';
-import { AbstractRunnableTestSuiteInfo } from '../AbstractRunnableTestSuiteInfo';
-import { AbstractTestInfo } from '../AbstractTestInfo';
-import { AbstractTestSuiteInfoBase } from '../AbstractTestSuiteInfoBase';
-import { DOCTestInfo } from './DOCTestInfo';
+import { RunnableSuiteProperties } from '../RunnableSuiteProperties';
+import { AbstractRunnableSuite } from '../AbstractRunnableSuite';
+import { AbstractSuit } from '../AbstractSuit';
+import { Catch2Test } from './Catch2Test';
 import { SharedVariables } from '../SharedVariables';
 import { RunningTestExecutableInfo, ProcessResult } from '../RunningTestExecutableInfo';
-import { RunnableTestSuiteProperties } from '../RunnableTestSuiteProperties';
-import { GroupTestSuiteInfo } from '../GroupTestSuiteInfo';
+import { AbstractTest } from '../AbstractTest';
+import { GroupSuite } from '../GroupSuite';
 
 interface XmlObject {
   [prop: string]: any; //eslint-disable-line
 }
 
-export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
+export class Catch2Suite extends AbstractRunnableSuite {
   public constructor(
     shared: SharedVariables,
     label: string,
     desciption: string | undefined,
-    execInfo: RunnableTestSuiteProperties,
-    docVersion: [number, number, number] | undefined,
+    execInfo: RunnableSuiteProperties,
+    catch2Version: [number, number, number] | undefined,
   ) {
-    super(shared, label, desciption, execInfo, 'doctest', Promise.resolve(docVersion));
+    super(shared, label, desciption, execInfo, 'Catch2', Promise.resolve(catch2Version));
   }
 
-  private _reloadFromString(
-    testListOutput: string,
-    oldChildren: (AbstractTestSuiteInfoBase | AbstractTestInfo)[],
-  ): void {
-    let res: XmlObject = {};
-    new xml2js.Parser({ explicitArray: true }).parseString(testListOutput, (err: Error, result: XmlObject) => {
-      if (err) {
-        throw err;
-      } else {
-        res = result;
+  private _reloadFromString(testListOutput: string, oldChildren: (AbstractSuit | AbstractTest)[]): void {
+    const lines = testListOutput.split(/\r?\n/);
+
+    const startRe = /Matching test cases:/;
+    const endRe = /[0-9]+ matching test cases?/;
+
+    let i = 0;
+
+    while (i < lines.length) {
+      const m = lines[i++].match(startRe);
+      if (m !== null) break;
+    }
+
+    if (i >= lines.length) {
+      this._shared.log.error('Wrong test list output format #1', testListOutput);
+      throw Error('Wrong test list output format');
+    }
+
+    while (i < lines.length) {
+      const m = lines[i].match(endRe);
+      if (m !== null) break;
+
+      if (!lines[i].startsWith('  ')) this._shared.log.error('Wrong test list output format', i, lines);
+
+      if (lines[i].startsWith('    ')) {
+        this._shared.log.warn('Probably too long test name', i, lines);
+        this.children = [];
+        const test = this.addChild(
+          new Catch2Test(this._shared, undefined, 'Check the test output message for details ⚠️', '', [], '', 0),
+        );
+        this._shared.sendTestEventEmitter.fire([
+          {
+            type: 'test',
+            test: test,
+            state: 'errored',
+            message: [
+              '⚠️ Probably too long test name or the test name starts with space characters!',
+              '🛠 - Try to define `CATCH_CONFIG_CONSOLE_WIDTH 300` before `catch2.hpp` is included.',
+              '🛠 - Remove whitespace characters from the beggining of test "' + lines[i].substr(2) + '"',
+            ].join('\n'),
+          },
+        ]);
+        return;
       }
-    });
+      const testNameAsId = lines[i++].substr(2);
 
-    for (let i = 0; i < res.doctest.TestCase.length; ++i) {
-      const testCase = res.doctest.TestCase[i].$;
+      let filePath: string | undefined = undefined;
+      let line: number | undefined = undefined;
+      {
+        const fileLine = lines[i++].substr(4);
+        const match = fileLine.match(/(?:(.+):([0-9]+)|(.+)\(([0-9]+)\))/);
 
-      const testNameAsId = testCase.name;
-      const filePath: string | undefined = testCase.filename ? this._findFilePath(testCase.filename) : undefined;
-      const line: number | undefined = testCase.line !== undefined ? Number(testCase.line) - 1 : undefined;
-      const skipped: boolean | undefined = testCase.skipped !== undefined ? testCase.skipped === 'true' : undefined;
-      const suite: string | undefined = testCase.testsuite !== undefined ? testCase.testsuite : undefined;
+        if (match && match.length == 5) {
+          const matchedPath = match[1] ? match[1] : match[3];
+          filePath = this._findFilePath(matchedPath);
+          line = Number(match[2] ? match[2] : match[4]) - 1;
+        } else {
+          this._shared.log.error('Could not find catch2 file info', lines);
+        }
+      }
+
+      let description = lines[i++].substr(4);
+      if (description.startsWith('(NO DESCRIPTION)')) description = '';
+
+      let tags: string[] = [];
+      if (i < lines.length && lines[i].length > 6 && lines[i][6] === '[') {
+        tags = lines[i].trim().split(']');
+        tags.pop();
+        for (let j = 0; j < tags.length; ++j) tags[j] += ']';
+        ++i;
+      }
 
       const old = this.findTestInfoInArray(oldChildren, v => v.testNameAsId === testNameAsId);
 
-      const test = new DOCTestInfo(
+      const test = new Catch2Test(
         this._shared,
-        undefined,
+        old ? old.id : undefined,
         testNameAsId,
-        suite !== undefined ? `[${suite}]` : undefined,
-        skipped,
+        description,
+        tags,
         filePath,
         line,
-        old as DOCTestInfo,
+        old ? (old as Catch2Test).sections : undefined,
       );
 
-      let group = this as AbstractTestSuiteInfoBase;
+      let group = this as AbstractSuit;
 
       if (this.execInfo.groupBySource && test.file) {
         const fileStr = pathlib.basename(test.file);
@@ -73,12 +123,28 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
           group = found;
         } else {
           const oldGroup = this.findGroupInArray(oldChildren, v => v.origLabel === fileStr);
-          group = group.addChild(new GroupTestSuiteInfo(this._shared, fileStr, oldGroup));
+          group = group.addChild(new GroupSuite(this._shared, fileStr, oldGroup));
+        }
+      }
+
+      if (this.execInfo.groupByTags && tags.length > 0) {
+        const tagsStr = tags
+          .filter(v => v != '[.]' && v != '[hide]')
+          .sort()
+          .join('');
+        const found = this.findGroup(v => v.origLabel === tagsStr);
+        if (found) {
+          group = found;
+        } else {
+          const oldGroup = this.findGroupInArray(oldChildren, v => v.origLabel === tagsStr);
+          group = group.addChild(new GroupSuite(this._shared, tagsStr, oldGroup));
         }
       }
 
       group.addChild(test);
     }
+
+    if (i >= lines.length) this._shared.log.error('Wrong test list output format #2', lines);
   }
 
   protected async _reloadChildren(): Promise<void> {
@@ -97,94 +163,91 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
           this._shared.log.info('loading from cache: ', cacheFile);
           const content = await promisify(fs.readFile)(cacheFile, 'utf8');
 
-          this._reloadFromString(content, oldChildren);
-          return Promise.resolve();
+          if (content === '') {
+            this._shared.log.debug('loading from cache failed because file is empty');
+          } else {
+            this._reloadFromString(content, oldChildren);
+
+            return;
+          }
         }
       } catch (e) {
         this._shared.log.warn('coudnt use cache', e);
       }
     }
 
-    return c2fs
-      .spawnAsync(
-        this.execInfo.path,
-        this.execInfo.prependTestListingArgs.concat([
-          '--list-test-cases',
-          '--reporters=xml',
-          '--no-skip=true',
-          '--no-color=true',
-        ]),
-        this.execInfo.options,
-        30000,
-      )
-      .then(docTestListOutput => {
-        if (docTestListOutput.stderr && !this.execInfo.ignoreTestEnumerationStdErr) {
-          this._shared.log.warn(
-            'reloadChildren -> docTestListOutput.stderr',
-            docTestListOutput.stdout,
-            docTestListOutput.stderr,
-            docTestListOutput.error,
-            docTestListOutput.status,
-          );
-          const test = this.addChild(
-            new DOCTestInfo(
-              this._shared,
-              undefined,
-              'Check the test output message for details ⚠️',
-              undefined,
-              false,
-              undefined,
-              undefined,
-            ),
-          );
-          this._shared.sendTestEventEmitter.fire([
-            {
-              type: 'test',
-              test: test,
-              state: 'errored',
-              message: [
-                `❗️Unexpected stderr!`,
-                `(One might can use ignoreTestEnumerationStdErr as the LAST RESORT. Check README for details.)`,
-                `spawn`,
-                `stout:`,
-                `${docTestListOutput.stdout}`,
-                `stderr:`,
-                `${docTestListOutput.stderr}`,
-              ].join('\n'),
-            },
-          ]);
-          return Promise.resolve();
-        }
+    const catch2TestListOutput = await c2fs.spawnAsync(
+      this.execInfo.path,
+      this.execInfo.prependTestListingArgs.concat([
+        '[.],*',
+        '--verbosity',
+        'high',
+        '--list-tests',
+        '--use-colour',
+        'no',
+      ]),
+      this.execInfo.options,
+      30000,
+    );
 
-        this._reloadFromString(docTestListOutput.stdout, oldChildren);
+    if (catch2TestListOutput.stderr && !this.execInfo.ignoreTestEnumerationStdErr) {
+      this._shared.log.warn('reloadChildren -> catch2TestListOutput.stderr', catch2TestListOutput);
+      const test = this.addChild(
+        new Catch2Test(this._shared, undefined, 'Check the test output message for details ⚠️', '', [], '', 0),
+      );
+      this._shared.sendTestEventEmitter.fire([
+        {
+          type: 'test',
+          test: test,
+          state: 'errored',
+          message: [
+            `❗️Unexpected stderr!`,
+            `(One might can use ignoreTestEnumerationStdErr as the LAST RESORT. Check README for details.)`,
+            `spawn`,
+            `stout:`,
+            `${catch2TestListOutput.stdout}`,
+            `stderr:`,
+            `${catch2TestListOutput.stderr}`,
+          ].join('\n'),
+        },
+      ]);
+      return Promise.resolve();
+    }
 
-        if (this._shared.enabledTestListCaching) {
-          return promisify(fs.writeFile)(cacheFile, docTestListOutput.stdout).catch(err =>
-            this._shared.log.warn('couldnt write cache file:', err),
-          );
-        }
-        return Promise.resolve();
-      });
+    if (catch2TestListOutput.stdout.length === 0) {
+      this._shared.log.debug(catch2TestListOutput);
+      throw Error('stoud is empty');
+    }
+
+    this._reloadFromString(catch2TestListOutput.stdout, oldChildren);
+
+    if (this._shared.enabledTestListCaching) {
+      return promisify(fs.writeFile)(cacheFile, catch2TestListOutput.stdout).catch(err =>
+        this._shared.log.warn('couldnt write cache file:', err),
+      );
+    }
   }
 
-  protected _getRunParams(childrenToRun: 'runAllTestsExceptSkipped' | Set<DOCTestInfo>): string[] {
+  protected _getRunParams(childrenToRun: 'runAllTestsExceptSkipped' | Set<Catch2Test>): string[] {
     const execParams: string[] = [];
 
     if (childrenToRun !== 'runAllTestsExceptSkipped') {
       const testNames = [...childrenToRun].map(c => c.getEscapedTestName());
-      execParams.push('--test-case=' + testNames.join(','));
-      execParams.push('--no-skip=true');
+      execParams.push(testNames.join(','));
     }
 
-    execParams.push('--case-sensitive=true');
-    execParams.push('--reporters=xml');
-    execParams.push('--duration=true');
+    execParams.push('--reporter');
+    execParams.push('xml');
+    execParams.push('--durations');
+    execParams.push('yes');
 
-    if (this._shared.isNoThrow) execParams.push('--no-throw=true');
+    if (this._shared.isNoThrow) execParams.push('--nothrow');
 
     if (this._shared.rngSeed !== null) {
-      execParams.push('--order-by=rand');
-      execParams.push('--rand-seed=' + this._shared.rngSeed.toString());
+      execParams.push('--order');
+      execParams.push('rand');
+      execParams.push('--rng-seed');
+      execParams.push(this._shared.rngSeed.toString());
     }
 
     return execParams;
@@ -194,15 +257,15 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
     const data = new (class {
       public buffer = '';
       public inTestCase = false;
-      public currentChild: AbstractTestInfo | undefined = undefined;
-      public route: AbstractTestSuiteInfoBase[] = [];
+      public currentChild: AbstractTest | undefined = undefined;
+      public route: AbstractSuit[] = [];
       public beforeFirstTestCase = true;
       public rngSeed: number | undefined = undefined;
       public unprocessedXmlTestCases: string[] = [];
-      public processedTestCases: AbstractTestInfo[] = [];
+      public processedTestCases: AbstractTest[] = [];
     })();
 
-    const testCaseTagRe = /<TestCase(\s+[^\n\r]+)[^\/](\/)?>/;
+    const testCaseTagRe = /<TestCase(?:\s+[^\n\r]+)?>/;
 
     return new Promise<ProcessResult>(resolve => {
       const chunks: string[] = [];
@@ -212,46 +275,40 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
         let invariant = 99999;
         do {
           if (!data.inTestCase) {
-            if (data.beforeFirstTestCase && data.rngSeed === undefined) {
-              const ri = data.buffer.match(/<Options\s+[^>\n]*rand_seed="([0-9]+)"/);
-              if (ri != null && ri.length == 2) {
-                data.rngSeed = Number(ri[1]);
-              }
-            }
+            const b = data.buffer.indexOf('<TestCase');
+            if (b == -1) return;
 
             const m = data.buffer.match(testCaseTagRe);
-            if (m == null) return;
+            if (m == null || m.length != 1) return;
 
-            const skipped = m[2] === '/';
             data.inTestCase = true;
-            let name = '';
 
-            if (skipped) {
-              new xml2js.Parser({ explicitArray: true }).parseString(m[0], (err: Error, result: XmlObject) => {
+            let name = '';
+            new xml2js.Parser({ explicitArray: true }).parseString(
+              m[0] + '</TestCase>',
+              (err: Error, result: XmlObject) => {
                 if (err) {
                   this._shared.log.exception(err);
                   throw err;
                 } else {
                   name = result.TestCase.$.name;
                 }
-              });
-            } else {
-              new xml2js.Parser({ explicitArray: true }).parseString(
-                m[0] + '</TestCase>',
-                (err: Error, result: XmlObject) => {
-                  if (err) {
-                    this._shared.log.exception(err);
-                    throw err;
-                  } else {
-                    name = result.TestCase.$.name;
-                  }
-                },
-              );
+              },
+            );
+
+            if (data.beforeFirstTestCase) {
+              const ri = data.buffer.match(/<Randomness\s+seed="([0-9]+)"\s*\/?>/);
+              if (ri != null && ri.length == 2) {
+                data.rngSeed = Number(ri[1]);
+              }
             }
 
             data.beforeFirstTestCase = false;
 
-            const [route, testInfo] = this.findRouteToTestInfo(v => v.testNameAsId == name);
+            const [route, testInfo] = this.findRouteToTestInfo(v => {
+              // xml output trimmes the name of the test
+              return v.testNameAsId.trim() == name;
+            });
 
             if (testInfo !== undefined) {
               this.sendMinimalEventsIfNeeded(data.route, route);
@@ -259,41 +316,15 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
 
               data.currentChild = testInfo;
               this._shared.log.info('Test', data.currentChild.testNameAsId, 'has started.');
-
-              if (!skipped) {
-                this._shared.testStatesEmitter.fire(data.currentChild.getStartEvent());
-                data.buffer = data.buffer.substr(m.index!);
-              } else {
-                this._shared.log.info('Test ', data.currentChild.testNameAsId, 'has skipped.');
-
-                // this always comes so we skip it
-                //const testCaseXml = m[0];
-                //this._shared.testStatesEmitter.fire(data.currentChild.getStartEvent());
-                // try {
-                //   const ev: TestEvent = data.currentChild.parseAndProcessTestCase(testCaseXml, data.rngSeed, runInfo);
-                //   data.processedTestCases.push(data.currentChild);
-                //   this._shared.testStatesEmitter.fire(ev);
-                // } catch (e) {
-                //   this._shared.log.error('parsing and processing test', e, data, testCaseXml);
-                //   this._shared.testStatesEmitter.fire({
-                //     type: 'test',
-                //     test: data.currentChild,
-                //     state: 'errored',
-                //     message: '😱 Unexpected error under parsing output !! Error: ' + inspect(e) + '\n',
-                //   });
-                // }
-
-                data.inTestCase = false;
-                data.currentChild = undefined;
-                data.buffer = data.buffer.substr(m.index! + m[0].length);
-              }
+              this._shared.testStatesEmitter.fire(data.currentChild.getStartEvent());
             } else {
               this._shared.log.info('TestCase not found in children', name);
             }
+
+            data.buffer = data.buffer.substr(b);
           } else {
             const endTestCase = '</TestCase>';
             const b = data.buffer.indexOf(endTestCase);
-
             if (b == -1) return;
 
             const testCaseXml = data.buffer.substring(0, b + endTestCase.length);
@@ -322,7 +353,7 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
                 });
               }
             } else {
-              this._shared.log.info('<TestCase> found without TestInfo: ', this, '; ', testCaseXml);
+              this._shared.log.info('<TestCase> found without TestInfo', this, testCaseXml);
               data.unprocessedXmlTestCases.push(testCaseXml);
             }
 
@@ -359,7 +390,7 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
 
         if (data.inTestCase) {
           if (data.currentChild !== undefined) {
-            this._shared.log.info('data.currentChild !== undefined: ', data);
+            this._shared.log.info('data.currentChild !== undefined', data);
             let ev: TestEvent;
 
             if (runInfo.timeout !== null) {
@@ -387,15 +418,14 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
         this.sendMinimalEventsIfNeeded(data.route, []);
         data.route = [];
 
-        // skipped test handling is wrong. I just disable this feature
-        // const isTestRemoved =
-        //   runInfo.timeout === null &&
-        //   result.error === undefined &&
-        //   ((runInfo.childrenToRun === 'runAllTestsExceptSkipped' &&
-        //     this.getTestInfoCount(false) > data.processedTestCases.length) ||
-        //     (runInfo.childrenToRun !== 'runAllTestsExceptSkipped' && data.processedTestCases.length == 0));
+        const isTestRemoved =
+          runInfo.timeout === null &&
+          result.error === undefined &&
+          ((runInfo.childrenToRun === 'runAllTestsExceptSkipped' &&
+            this.getTestInfoCount(false) > data.processedTestCases.length) ||
+            (runInfo.childrenToRun !== 'runAllTestsExceptSkipped' && data.processedTestCases.length == 0));
 
-        if (data.unprocessedXmlTestCases.length > 0 /*|| isTestRemoved*/) {
+        if (data.unprocessedXmlTestCases.length > 0 || isTestRemoved) {
           new Promise<void>((resolve, reject) => {
             this._shared.loadWithTaskEmitter.fire(() => {
               return this.reloadTests(this._shared.taskPool).then(resolve, reject);
@@ -426,7 +456,7 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
                 if (name === undefined) break;
 
                 // xml output trimmes the name of the test
-                const currentChild = this.findTestInfo(v => v.testNameAsId === name);
+                const currentChild = this.findTestInfo(v => v.testNameAsId.trim() == name);
 
                 if (currentChild === undefined) break;
 
@@ -437,7 +467,6 @@ export class DOCTestSuiteInfo extends AbstractRunnableTestSuiteInfo {
                   this._shared.log.error('parsing and processing test', e, testCaseXml);
                 }
               }
-
               events.length && this._shared.sendTestEventEmitter.fire(events);
             },
             (reason: Error) => {
