+++ src/framework/GoogleTest.ts
@@ -1,10 +1,10 @@
 import { TestEvent, TestDecoration } from 'vscode-test-adapter-api';
 
-import { AbstractTestInfo } from '../AbstractTestInfo';
+import { AbstractTest } from '../AbstractTest';
 import { SharedVariables } from '../SharedVariables';
 import { RunningTestExecutableInfo } from '../RunningTestExecutableInfo';
 
-export class GoogleTestInfo extends AbstractTestInfo {
+export class GoogleTest extends AbstractTest {
   public constructor(
     shared: SharedVariables,
     id: string | undefined,
